Fix timer tracking during suspend
Previously a monotonic timer was used to wake up the event loop whenever
an alarm should be rung. This would cause problems when suspending the
phone since it would cause the next alarm's timer to be delayed.

To fix this, the clock source was switched to a realtime clock based on
the tokio-walltime crate.