Fix RTC scheduling with elapsed wkalm set
This fixes an issue where Rezz would never schedule an alarm again when
a current RTC wakeup was set with an already elapsed time.

Instead rezz now only skips setting the alarm when there is already a
wakeup scheduled that is sooner than the current one, but past the
current time.